@@ -11,7 +11,7 @@
 	<description>Backend for SITU-WebApp</description>
 
 	<properties>
-		<java.version>20</java.version>
+		<java.version>21</java.version>
 	</properties>
 
 	<parent>
