@@ -10,7 +10,6 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 
-
 @RestController
 @RequestMapping("/api/situ/users")
 public class UserController {
@@ -21,14 +20,14 @@ public class UserController {
         this.userService = userService;
     }
 
-    @GetMapping("/list/{pageIndex}/{pageSize}/{companyId}")
+    @GetMapping("/{pageIndex}/{pageSize}/{companyId}")
     public Page<User> list(@PathVariable("pageIndex") Integer pageIndex,
                            @PathVariable("pageSize") Integer pageSize,
                            @PathVariable("companyId") Long companyId) {
         return this.userService.listUsers(pageIndex, pageSize, companyId);
     }
 
-    @GetMapping("/get/{id}")
+    @GetMapping("/{id}")
     public ResponseEntity<User> get(@PathVariable("id") Long id) {
         User user = this.userService.getUser(id);
         if (user == null) {
@@ -37,14 +36,12 @@ public class UserController {
         return ResponseEntity.ok(user);
     }
 
-    @PostMapping("/create")
+    @PostMapping()
     public ResponseEntity<Long> create(@RequestBody User user) {
         User userCreated = this.userService.createUser(user);
         if (userCreated == null) {
             return ResponseEntity.status(HttpStatus.NOT_FOUND).body(null);
         }
-
-        // return the user id
         return ResponseEntity.ok(userCreated.getId());
     }
 }
